+++ src/app/billing/send-quote/page.tsx
@@ -1,4 +1,4 @@
-import ContentContainer from '@/components/ui/containers/content-container';
+import FillFormContainer from '@/components/ui/containers/fill-form-container';
 import FormContainer from '@/components/ui/containers/form-container';
 import FormHeader from '@/components/ui/header/form-header';
 import { CreateQuoteForm } from '@/components/ui/stripe-quote-form/create-quote-form';
@@ -6,12 +6,12 @@ import { CreateQuoteForm } from '@/components/ui/stripe-quote-form/create-quote-
 export default function SendQuotePage() {
   return (
     <FormContainer>
-      <ContentContainer>
+      <FillFormContainer>
         <FormHeader text={'Create Stripe Quote'} />
         <div className="p-4 border rounded-md shadow-sm">
           <CreateQuoteForm />
         </div>
-      </ContentContainer >
+      </FillFormContainer>
     </FormContainer >
   );
 }
