+++ src/app/email/individual/page.tsx
@@ -1,7 +1,7 @@
 import SendEmailInputs from "@/components/ui/emails/send-email-inputs";
 import SendIndividualEmail from "@/components/ui/emails/send-individual-email";
 import SendIndividualEmailFallback from "@/components/ui/fallbacks/send-individual-email-fallback";
-import { fetchClientsNamesAndEmails } from "@/DAL/dal";
+import { fetchClientsNamesAndEmails } from "@/DAL/dal-clients";
 import { Suspense } from "react";
 
 export default function page() {
