+++ src/app/lists/clearing/page.tsx
@@ -1,15 +1,14 @@
 import SearchForm from "@/components/ui/client-list/search-form";
 import FormContainer from "@/components/ui/containers/form-container";
-// import ClientListService from "@/components/ui/service-list/clients-list-service";
 import FormHeader from "@/components/ui/header/form-header";
 import ClientListService from "../../../components/ui/service-list/clients-list-service";
-import { fetchSnowClearingClients } from "@/DAL/dal";
+import { fetchSnowClearingClients } from "@/DAL/dal-clients";
 import { isOrgAdmin } from "@/lib/webhooks";
 import { redirect } from "next/navigation";
 import { Suspense } from "react";
-// import { fetchOrgMembers } from "@/DAL/dal";
 import { parseClientListParams } from "@/lib/params";
-import { SearchParams } from "@/types/types";
+import { SearchParams } from "@/types/types-params";
+
 
 export default async function page({ searchParams }: { searchParams: Promise<SearchParams> }) {
     const [{ isAdmin, userId }, params] = await Promise.all([
