+++ src/app/lists/cutting/page.tsx
@@ -1,13 +1,13 @@
 import SearchForm from "@/components/ui/client-list/search-form";
 import FormContainer from "@/components/ui/containers/form-container";
 import FormHeader from "@/components/ui/header/form-header";
-import { fetchCuttingClients } from "@/DAL/dal";
+import { fetchCuttingClients } from "@/DAL/dal-clients";
 import { isOrgAdmin } from "@/lib/webhooks";
 import { redirect } from "next/navigation";
 import { Suspense } from "react";
 import ClientListService from "../../../components/ui/service-list/clients-list-service";
 import { parseClientListParams } from "@/lib/params";
-import { SearchParams } from "@/types/types";
+import { SearchParams } from "@/types/types-params";
 
 export default async function page({ searchParams }: { searchParams: Promise<SearchParams>; }) {
     const [{ isAdmin }, params] = await Promise.all([
