+++ src/components/ui/map-component/many-points-map.tsx
@@ -3,7 +3,7 @@ import Image from 'next/image';
 import Link from 'next/link';
 import FormHeader from '../header/form-header';
 import { useGetLocation, useGetLonAndLatFromAddresses } from '@/hooks/hooks';
-import { GeocodeResult, MapComponentProps } from '@/types/types';
+import { GeocodeResult, MapComponentProps } from '@/types/types-google-map-iframe';
 
 export default function ManyPointsMap({ addresses }: MapComponentProps) {
 
@@ -31,7 +31,7 @@ export default function ManyPointsMap({ addresses }: MapComponentProps) {
     })
     .join('&');
   const userMarker = `markers=size:mid%7Ccolor:blue%7C${userLocation.lat},${userLocation.lng}`;
-
+  //TODO: fix type issue here on . at and .log
   const mapUrl = `https://maps.googleapis.com/maps/api/staticmap?center=${center.lat},${center.lng}&zoom=&size=500x200&maptype=roadmap&${userMarker}&${markers}&key=${process.env.NEXT_PUBLIC_REACT_APP_GOOGLE_MAPS_API_KEY}`;
 
   const origin = `${userLocation.lat},${userLocation.lng}`;
