+++ src/components/ui/nav/nav-bar.tsx
@@ -1,12 +1,13 @@
 "use client"
 import * as React from "react"
 import { Button } from "../button"
-import { useState } from "react"
+import { Suspense, useState } from "react"
 import { Menu } from "lucide-react"
 
 import NavigationMenuComponent from "./nav-menu-component"
 
-export function NavBar({ userId, isAdmin }: { userId: string, isAdmin: boolean }) {
+export function NavBar({ userId, isAdmin, hasStripAPIKeyPromise }: { userId: string, isAdmin: boolean, hasStripAPIKeyPromise: Promise<boolean> }) {
+
     const [showNav, setShowNav] = useState(false)
 
     return (
@@ -15,10 +16,10 @@ export function NavBar({ userId, isAdmin }: { userId: string, isAdmin: boolean }
                 <Button variant="outline" onClick={() => setShowNav(!showNav)} className="p-2 w-9">
                     <Menu size={24} />
                 </Button>
-                {showNav && <NavigationMenuComponent userId={userId} isAdmin={isAdmin} />}
+                {showNav && <Suspense><NavigationMenuComponent hasStripAPIKeyPromise={hasStripAPIKeyPromise} userId={userId} isAdmin={isAdmin} /></Suspense>}
             </div>
 
-            <div className="hidden md:block"><NavigationMenuComponent userId={userId} isAdmin={isAdmin} /></div>
+            <div className="hidden md:block"><Suspense><NavigationMenuComponent hasStripAPIKeyPromise={hasStripAPIKeyPromise} userId={userId} isAdmin={isAdmin} /></Suspense></div>
         </>
     )
 }
