+++ src/components/ui/settings/price-per-cut-input.tsx
@@ -1,4 +1,4 @@
-import { Price } from "@/types/types";
+import { Price } from "@/types/types-clients";
 import { InputDiv } from "../containers/input-dev";
 import { InputField } from "../inputs/input";
 import { SettingsLabel } from "./settings-label";
@@ -8,7 +8,7 @@ export default async function PricePerCutInput({ pricePerCutPromise }: { pricePe
     return (
         <InputDiv >
             <SettingsLabel text="Price Per Cut" />
-            <InputField id={"price_pre_cut"} name={"price_pre_cut"} type={"number"} step="0.01" defaultValue={pricePerCut || 0.00} />
+            <InputField name={"price_pre_cut"} type={"number"} step={0.01} defaultValue={pricePerCut || 0.00} />
         </InputDiv>
     );
 }
