+++ src/types/types-components.ts
@@ -1,0 +1,22 @@
+import { Client } from "./types-clients";
+
+export interface CuttingPeriodSelectorProps {
+    label: string;
+    options: { value: string; label: string }[];
+    value: string;
+    handleChange: (event: React.ChangeEvent<HTMLSelectElement>) => void;
+    name: string;
+}
+
+export interface MaterialField {
+    id: string;
+    materialType?: string;
+    materialCostPerUnit?: number;
+    materialUnits?: number;
+}
+
+export interface ImageGalleryProps {
+  isAdmin: boolean;
+  client: Client;
+  setView: (view: string) => void;
+}
