+++ src/types/types-params.ts
@@ -1,0 +1,17 @@
+export type SearchParams = Record<string, string | string[] | number | undefined>;
+
+export interface ParsedClientListParams {
+    page: number;
+    searchTerm: string;
+    serviceDate: Date;
+    searchTermIsServiced: boolean;
+    searchTermCuttingWeek: number;
+    searchTermCuttingDay: string;
+    searchTermAssignedTo: string;
+    searchTermStatus: string;
+}
+
+export interface Props {
+    showForm: boolean;
+    setShowForm: React.Dispatch<React.SetStateAction<boolean>>;
+}
