+++ src/types/types-resend.ts
@@ -1,0 +1,7 @@
+export interface HeaderEmailProps {
+    text: string;
+    senderName: string;
+    companyName: string;
+    title: string;
+}
+
