+++ src/types/types-site-map.ts
@@ -1,0 +1,8 @@
+import { Client } from "./types-clients";
+
+export interface ImageListViewProps {
+  client: Client;
+  setView: (view: string) => void;
+  previewSrc: string | null;
+  setPreviewSrc: (src: string | null) => void;
+}
